Update to ubuntu 13.10.
There were some issues with this upgrade.

By far the most prominent was fixing some compatability changes for
boost. See hhvm commit fee428f for details - this commit was backported
to the version of hhvm used for hphpc. This fixed issues where
boost::shared_ptr now has explicit operator bool, an issue with the use
of shared_ptr<T[]> in debugger_client.h which required replacing it with
a shared_array<T>, and adding a null_vertex function to
ControlFlowGraph.

std::has_trivial_destructor was renamed to
std::is_trivially_destructable. This required some fixes, particularly
in the folly library.

The version of boost also moved the definition of boost::lock_guard to
a separate header, which required a fix to folly.

The version of GD in ubuntu 13.10 provides its own gdImageConvolution
and also removed the gdGetC function. Fix was to remove the copy
and pasted gdImageConvolution from ext_image.cpp and to copy the
(relatively trivial definition) of gdGetC in.

-Wno-unused-local-typedefs was also added to the build flags, because a
new proliferation of unused-local-typedef warnings was causing
difficulty in analyzing build breakages.